http: simplify buffer size limiting
We limit the `parse_buffer` that we use to fill data from the network to
the size of the output buffer from the caller.  Update this to
(temporarily) shrink the length field, instead of trying to manipulate
how filled the buffer is (in the `offset` field).  Manipulating the
size of the buffer is simpler and allows us to hold the used length
(`offset`) unchanged.